backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
groq==0.9.0
networkx==3.3
pandas==2.2.0
python-dotenv==1.0.0
httpx==0.27.2
